Assistant checkpoint: Added shopping cart functionality with new files and styles
Assistant generated file changes:
- cart.html: Create cart page
- cart.css: Add cart styles
- cart.js: Add cart functionality
- index.html: Add cart icon to header, Add cart.css and cart.js to index, Add cart.js script to index

---

User prompt:

can we add a cart page that will atuomatically add up your total from what you ordered and make it visitable with a little cart logo in the top right. Also ake a ittle number pop up on the top right of the cart with the number of items you ordered. Please make this a seperate file or files if possible.